exploring-ia-hugging-face/requirements.txt
transformers==4.47.1
python-dotenv==1.1.1